requirements.txt
aiohttp
beautifulsoup4
click
pytest
python-dotenv
PyYAML
requests
ruff
xmltodict
